Update p_tost_converter.R
I've changed the code to include special cases (cf. Lucy McGowan):
- when H0 = point H0
- when there is no CI around I (ie CI = point CI)

Few comments: 
1) when H0 = point H0 --> res$TOST_p1+res$TOST_p2 = 1. Because we always return max(res$TOST_p1,res$TOST_p2), p_tost cannot be < .5. If a user introduces p_tost < .5, there is an error message (the visual is not very good, we can probably improve it)

2) Point CI is mathematically possible only if ...
... n=infinite (impossible in practice)
... sd  = 0 

If SD = 0, data are constant (>< variable). And therefore, the normality assumption cannot be met. Moreover, t = infinite (because the denominator equals 0), and it is not possible to compute p_tost. If a user introduces sd=0, there is a message to explain that SGPV is computed based on I=m-mu and the eqbounds (low and high; not based on p_tost).

3) In the previous code, when CI = non-point CI and HO = non-point H0, I used these two lines below:
 if (lb > high_eqbound | ub < low_eqbound){SGPV = 0} 
 else if(lb > low_eqbound & ub < high_eqbound){SGPV = 1} 

I replaced it by ...
        if (p_tost>(1-alpha/2)){SGPV = 0} 
        else if(p_tost<(alpha/2)){SGPV = 1} 

==> It is more consistent with the idea to convert p_tost into SGPV. But there is a disadvantage: if users returns the wrong p from TOSTER (e.g. res$TOST_p1 instead of res$TOST_p2), the functions returns the wrong SGPV (I think it returns 1-SGPV).